+++ Source/GA6thFinal_Framework/AudioModule/AudioModule.vcxproj
@@ -140,20 +140,24 @@
     <ClInclude Include="AudioModule.h" />
     <ClInclude Include="Exceptions.h" />
     <ClInclude Include="Handle.h" />
+    <ClInclude Include="Callback.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="Requires.h" />
+    <ClInclude Include="Source.h" />
     <ClInclude Include="System.h" />
     <ClInclude Include="Types.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Exceptions.cpp" />
     <ClCompile Include="Handle.cpp" />
+    <ClCompile Include="Callback.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="Source.cpp" />
     <ClCompile Include="System.cpp" />
     <ClCompile Include="Types.cpp" />
   </ItemGroup>
