+++ Source/GA6thFinal_Framework/AudioModule/Callback.cpp
@@ -1,0 +1,13 @@
+﻿#include "pch.h"
+#include "Callback.h"
+
+namespace Audio
+{
+    Callback::Callback() = default;
+
+    void Callback::OnBufferEnd(void* pBufferContext)
+    {
+        if (_onBufferEnd)
+            _onBufferEnd();
+    }
+} // namespace Audio
