+++ Source/GA6thFinal_Framework/AudioModule/Requires.h
@@ -4,6 +4,13 @@
 #include <string>
 #include <filesystem>
 #include <fstream>
+#include <unordered_map>
+#include <functional>
+
+#include <xaudio2.h>
+#include <winrt/base.h>
 
 #include "Exceptions.h"
 #include "Types.h"
+#include "Handle.h"
+#include "Source.h"
