+++ Source/GA6thFinal_Framework/GameEngine/Source/UmFramework.h
@@ -154,11 +154,7 @@ using namespace Microsoft::WRL;
 #include "Engine/GraphicsCore/GraphicsCore.h"
 
 //Audio Core
-#include "Engine/AudioCore/Declare/Types/AudioTypes.h"
-#include "Engine/AudioCore/Declare/Exceptions/AudioExceptions.h"
-#include "Engine/AudioCore/Handle/AudioHandle.h"
-#include "Engine/AudioCore/Declare/Callback/IncreaseGenerationCallback.h"
-#include "Engine/AudioCore/AudioManager/AudioManager.h"
+#include "../AudioModule/AudioModule.h"
 
 //Engine Core
 #include "Engine/EngineCore/EngineLogger.h"
